OCTOBER/10-26/GCSplit1-10-26/requirements.txt
Flask==3.0.0
google-cloud-secret-manager==2.16.4
google-cloud-sql-connector[pg8000]==1.4.3
google-cloud-tasks==2.14.2
pg8000==1.30.3
